Rewrite Python example
