--- python/resolv.conf
@@ -1,2 +1,0 @@
-nameserver 8.8.8.8
-nameserver 8.8.4.4
