--- python/scripts/helloworld.py
@@ -1,3 +1,0 @@
-#!/usr/bin/env python3
-
-print("Hello World")
